+++ packages/gatsby-recipes/src/find-dependency-match.js
@@ -1,29 +1,38 @@
-module.exports = (allResources, r) => {
-  if (!r.dependsOn) {
+// Match a resource's defined dependencies against all resources
+// defined in the recipe.
+module.exports = (resources, resourceWithDependencies) => {
+  // Resource doesn't have a dependsOn key so we return
+  if (!resourceWithDependencies.dependsOn) {
     return []
   } else {
-    return r.dependsOn.map(d => {
-      const { resourceName, ...otherValues } = d
+    return resourceWithDependencies.dependsOn.map(dependency => {
+      const { resourceName, ...otherValues } = dependency
       const keys = Object.keys(otherValues)
 
-      const match = allResources.find(r1 => {
+      const match = resources.find(resource => {
         // Is this the right resourceName?
-        if (resourceName !== r1.resourceName) {
+        if (resourceName !== resource.resourceName) {
           return false
         }
         // Do keys match?
         if (
-          !keys.every(k => Object.keys(r1.resourceDefinitions).indexOf(k) >= 0)
+          !keys.every(
+            key => Object.keys(resource.resourceDefinitions).indexOf(key) >= 0
+          )
         ) {
           return false
         }
 
         // Do values match?
-        if (!keys.every(k => r1.resourceDefinitions[k] === d[k])) {
+        if (
+          !keys.every(
+            key => resource.resourceDefinitions[key] === dependency[key]
+          )
+        ) {
           return false
         }
 
-        return r1
+        return resource
       })
 
       if (match) {
@@ -33,12 +42,16 @@ module.exports = (allResources, r) => {
           // eslint-disable-next-line
           mdxType,
           ...resourceDefinition
-        } = r.resourceDefinitions
+        } = resourceWithDependencies.resourceDefinitions
 
         return {
-          error: `A resource (${r.resourceName}: ${JSON.stringify(
+          error: `A resource (${
+            resourceWithDependencies.resourceName
+          }: ${JSON.stringify(
             resourceDefinition
-          )}) is missing its dependency on ${JSON.stringify(r.dependsOn[0])}`,
+          )}) is missing its dependency on ${JSON.stringify(
+            resourceWithDependencies.dependsOn[0]
+          )}`,
         }
       }
     })