+++ packages/gatsby-recipes/src/recipe-machine/errors.test.js
@@ -36,10 +36,10 @@ describe(`recipe-machine errors`, () => {
   it(`errors if the introduction step has a command`, done => {
     const initialContext = {
       src: `
-# Hello, world
+  # Hello, world
 
-<File path="./hi.md" content="#yo" />
-    `,
+  <File path="./hi.md" content="#yo" />
+      `,
     }
     const service = interpret(
       recipeMachine.withContext(initialContext)