+++ packages/gatsby-recipes/src/recipe-machine/index.js
@@ -2,10 +2,9 @@ const { Machine, assign, send } = require(`xstate`)
 const lodash = require(`lodash`)
 
 const debug = require(`debug`)(`recipes-machine`)
-
+const validateSteps = require(`../validate-steps`)
 const createPlan = require(`../create-plan`)
 const applyPlan = require(`../apply-plan`)
-const validateSteps = require(`../validate-steps`)
 const parser = require(`../parser`)
 const resolveRecipe = require(`../resolve-recipe`)
 const findDependencyMatch = require(`../find-dependency-match`)
@@ -20,6 +19,7 @@ const recipeMachine = Machine(
       recipe: ``,
       recipeSrc: ``,
       stepsAsMdx: [],
+      stepsAsJS: [],
       exports: [],
       plan: [],
       commands: [],
@@ -69,7 +69,6 @@ const recipeMachine = Machine(
             debug(`parsingRecipe`)
             const parsed = await parser.parse(context.recipeSrc)
             debug(`parsedRecipe`)
-
             return parsed
           },
           onError: {
@@ -96,6 +95,7 @@ const recipeMachine = Machine(
             actions: assign({
               recipe: (context, event) => event.data.recipe,
               stepsAsMdx: (context, event) => event.data.stepsAsMdx,
+              stepsAsJS: (context, event) => event.data.stepsAsJS,
               exports: (context, event) => event.data.exports,
             }),
           },
@@ -130,7 +130,7 @@ const recipeMachine = Machine(
           src: (context, event) => async (cb, onReceive) => {
             try {
               let result = await createPlan(context, cb)
-              // Validate dependencies are met in the plan
+              // Validate dependencies are met in the resources plan
               result = result.map(r => {
                 const matches = findDependencyMatch(result, r)
                 // If there's any errors, replace the resource