+++ packages/gatsby/src/redux/__tests__/jobsv2.js
@@ -1,7 +1,11 @@
 const jobsManager = require(`../../utils/jobs-manager`)
 jest.spyOn(jobsManager, `enqueueJob`)
 jest.spyOn(jobsManager, `removeInProgressJob`)
-jest.mock(`uuid/v4`, () => () => `1234`)
+jest.mock(`uuid`, () => {
+  return {
+    v4: () => `1234`,
+  }
+})
 
 import { jobsV2Reducer as jobsReducer } from "../reducers/jobsv2"