+++ packages/gatsby/src/services/types.ts
@@ -11,13 +11,13 @@ import { Compiler } from "webpack"
 import { WebsocketManager } from "../utils/websocket-manager"
 import { IWebpackWatchingPauseResume } from "../utils/start-server"
 export interface IGroupedQueryIds {
-  pageQueryIds: string[]
-  staticQueryIds: string[]
+  pageQueryIds: Array<string>
+  staticQueryIds: Array<string>
 }
 
 export interface IMutationAction {
   type: string
-  payload: unknown[]
+  payload: Array<unknown>
   resolve?: (result: unknown) => void
 }
 export interface IBuildContext {
@@ -33,7 +33,7 @@ export interface IBuildContext {
   app?: Express
   nodesMutatedDuringQueryRun?: boolean
   mutationListener?: Actor<unknown, AnyEventObject>
-  nodeMutationBatch?: IMutationAction[]
+  nodeMutationBatch?: Array<IMutationAction>
   compiler?: Compiler
   websocketManager?: WebsocketManager
   webpackWatching?: IWebpackWatchingPauseResume