+++ www/gatsby-node.js
@@ -1,4 +1,5 @@
 const Promise = require(`bluebird`)
+const fs = require(`fs`)
 const startersRedirects = require(`./starter-redirects.json`)
 
 const { loadYaml } = require(`./src/utils/load-yaml`)
@@ -150,14 +151,14 @@ exports.createPages = async helpers => {
   // splat redirects
   await createRedirect({
     fromPath: `/packages/*`,
-    toPath: `https://gatsbyjs.com/plugins/:splat`,
+    toPath: `https://www.gatsbyjs.com/plugins/:splat`,
     isPermanent: true,
     force: true,
   })
 
   await createRedirect({
     fromPath: `/creators/*`,
-    toPath: `https://gatsbyjs.com/partner/`,
+    toPath: `https://www.gatsbyjs.com/partner/`,
     isPermanent: true,
     force: true,
   })
@@ -166,8 +167,16 @@ exports.createPages = async helpers => {
   //  this needs to be the last redirect created or it'll match everything
   await createRedirect({
     fromPath: `/*`,
-    toPath: `https://gatsbyjs.com/:splat`,
+    toPath: `https://www.gatsbyjs.com/:splat`,
     isPermanent: true,
     force: true,
   })
+}
+
+exports.onPostBuild = () => {
+  const redirectsString = fs.readFileSync(`./public/_redirects`).toString()
+  const chunked = redirectsString.split(`\n`)
+
+  chunked.splice(chunked.length - 1, 0, `/sw.js  /sw.js  200`)
+  fs.writeFileSync(`./public/_redirects`, chunked.join(`\n`))
 }