+++ packages/gatsby-plugin-preact/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.9...gatsby-plugin-preact@4.0.10) (2020-08-28)
+
+**Note:** Version bump only for package gatsby-plugin-preact
+
 ## [4.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.8...gatsby-plugin-preact@4.0.9) (2020-07-09)
 
 **Note:** Version bump only for package gatsby-plugin-preact